install moontoast/math v1.1.2
uuid need moontoast/math, but error happen when deploy to heroku

Here is the Heroku support response :

The issue here is your update of moontoast/math from 1.1.2 to 1.2.1

1.1.2 required ext-bcmath: https://packagist.org/packages/moontoast/math#1.1.2

1.2.1 uses phpseclib/bcmath_compat instead: https://packagist.org/packages/moontoast/math#1.2.1

The issue that follows then is that phpseclib/bcmath_compat via its composer.json metadata provides the ext-bcmath extension, but that extension is marked as replaced (because it's bundled with PHP) by PHP 7.3.13.

A replace is a provide plus a conflict, and that's why PHP 7.3.13 (with ext-bcmath available as a shared extension) conflicts.

This doesn't happen on your local machine because dependency resolution works differently there - your regular composer install step is working, but on Heroku, what's failing is the "platform install" step that installs PHP and all necessary extensions, not userland dependencies, using a special custom repository and installer, and a composer.json generated from your composer.lock.

I don't think there is a way to work around this, short of you transitioning to e.g. brick/math (moontoast/math is marked as abandoned), or downgrading to moontoast/math 1.1.2. I'll put it on the list of things to investigate on our side, but it's a very fundamental aspect of how Composer handles dependencies, and how we have to resolve app requirements, that make it an edge case we'll not likely be able to fix.

Hope this helps, and sorry I don't have happier news.